Make installation easier
This particular change adds a pyproject.toml file and does a minimal
amount of refactoring to make `app.py` into a CLI driveable application.
The application is installed as `tiktok_repost_downloader`.

Remove all of the ad hoc dependency installation logic and rely on
setuptools-scm to do the heavy lifting to install all of the
dependencies in order to use the project.

Add python 3.10 compatible type hints to all functions while here.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>